requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
boto3==1.35.2
botocore==1.35.2
certifi==2024.7.4
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.3.2
constantly==23.10.4
contourpy==1.2.1
cryptography==44.0.2
cycler==0.12.1
daphne==4.1.2
distlib==0.3.8
Django==5.1
django-ckeditor==6.7.3
django-js-asset==3.1.2
django-taggit==6.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
docutils==0.21.2
emoji==2.12.1
et_xmlfile==2.0.0
filelock==3.15.4
fonttools==4.53.1
gunicorn==23.0.0
holidays==0.79
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
jmespath==1.0.1
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
matplotlib==3.9.2
msgpack==1.1.0
numpy==2.0.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pillow==10.4.0
platformdirs==4.2.2
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydroid==0.1.4
pygame==2.6.0
Pygments==2.18.0
pyOpenSSL==25.0.0
pyparsing==3.1.2
pypiwin32==223
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
pywin32==306
PyYAML==6.0.2
redis==5.2.1
reportlab==4.3.1
requests==2.32.3
s3transfer==0.10.2
seaborn==0.13.2
service-identity==24.2.0
setuptools==78.1.0
six==1.16.0
sqlparse==0.5.1
tk==0.1.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.2.2
user-agents==2.2.0
virtualenv==20.26.3
whitenoise==6.9.0
zope.interface==7.2
